Seperate get_env function and Fix unset
