+++ src/ft_unset.c
@@ -6,7 +6,7 @@
 /*   By: gpaeng <gpaeng@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 16:49:34 by gpaeng            #+#    #+#             */
-/*   Updated: 2022/01/06 11:29:52 by gpaeng           ###   ########.fr       */
+/*   Updated: 2022/01/06 12:48:53 by gpaeng           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ void	ft_remove_env(int env_line)
 	jop.envp = env;
 }
 
-void	ft_check_args_unset(char c)
+int	ft_check_args_unset(char c)
 {
 	if (c == '_' || ft_isalpha(c) || ft_isdigit(c))
 		return (1);
@@ -71,7 +71,7 @@ void	ft_unset(char *args[])
 	while (args[i])
 	{
 		env_line = ft_check_env(args[i]);
-		if (ft_check_args_unset(args[i]))
+		if (ft_check_args_unset(*args[i]))
 		{
 			printf("minishell: unset: %s not a valid identifier", args[i]);
             return ;
